requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.4
django-environ==0.9.0
django-import-export==3.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
et-xmlfile==1.1.0
httpie==3.2.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
multidict==6.0.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==22.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
PySocks==1.7.1
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
xlrd==2.0.1
xlwt==1.3.0
